Better detect suitable libctf version
On some el9 distros, the version of libctf installed might not have
all the necessary features for the libabigail CTF reader, leading to
compilation errors due to missing types from the ctf-api.h header
file.  For instance, the ctf-api.h on some of those distros lacks the
definition of the type struct ctf_dict_t.

This patch adds a configure test for that struct and disables the CTF
support if that type is absent.

	* configure.ac: If the "struct ctf_dict_t" type is not present in
	the version of ctf-api.h that is present, then switch the support
	of CTF off.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>