ini: Fix parsing list property values
While looking at something else, I came across an issue in
read_context::read_list_property_value.  This function requires
elements of a list property value (separated by a comma) to be on a
single line, for instance.  This is because the code forgets to parse
white spaces after the comma.

Fixed thus.

	* src/abg-ini.cc (read_context::read_list_property_value): Expect
	white spaces after the comma.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>