{dwarf,elf_based}-reader,writer: Avoid duplicating corpora in corpus_group
It's been brought to my attention on IRC that running

    abidw --linux-tree <kernel-build-tree>

would result in a corpus group that duplicates every single corpus in
the resulting abixml.  Oops.

This is because both dwarf::reader::read_corpus() and
elf_based_reader::read_and_add_corpus_to_group() add the corpus to the
corpus_group, and yet, the later function calls the former.  So the
corpus is added to the corpus_group twice.

This patch ensures that
elf_based_reader::read_and_add_corpus_to_group() is the only one to
add the corpus to the group.  It also ensures that this happens before
the corpus is constructed from the debug info because that is useful
for sharing types among the various corpora.  Otherwise, those types
are potentially duplicated in the IR of each corpus.

The patch also ensures the abixml writer enforces the fact that each
corpus is emitted only once.

	* src/abg-dwarf-reader.cc (reader::read_debug_info_into_corpus):
	Do not add the corpus to the group here ...
	* src/abg-elf-based-reader.cc
	(elf_based_reader::read_and_add_corpus_to_group): ... because it's
	already added here.  But then, let's add it here /before/ reading
	type & symbols information into the corpus.
	* src/abg-writer.cc (write_context::m_emitted_corpora_set): Add
	new data member.
	(write_context::{corpus_is_emitted, record_corpus_as_emitted}):
	Define new member functions.
	(write_corpus): Invoke the new
	write_context::record_corpus_as_emitted here.
	(write_corpus_group): Ensure that each corpus is emitted only
	once.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>