tools-utils: Support kernel stablelist
Up until now, a kernel whitelist was expected to be a ini file with a
section having a name ending with the word "whitelist".  Nowadays,
they are called "stablelist", so the name of the section ends up with
"stablelist".  This patch makes
gen_suppr_spec_from_kernel_abi_whitelists support that.

	* src/abg-tools-utils.cc
	(gen_suppr_spec_from_kernel_abi_whitelists): Support section name
	that ends with the word 'stablelist'.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>