Bug 29911 - fedabipkgdiff forgets to provide some debuginfo RPMs to abipkgdiff
Some packages like foo-utils.rpm can be associated with two debuginfo
RPMs: One foo-utils-debuginfo.rpm and foo-debuginfo.rpm.  This is
because the foo-debuginfo.rpm contains debug info that has been
factorized out of all the sub-packages of foo, foo-utils being one of
those sub-packages.  In those cases, fedabipkgdiff needs to provide
foo-debuginfo.rpm and foo-utils-debuginfo.rpm to abipkgdiff so that it
can find all the necessary debuginfo.

This patch fixes fedabipkgdiff accordingly and adds some more logging
to abipkgdiff to make it emit an explicit message for cases like this.

	* tools/abipkgdiff.cc (compare_to_self): Emit an error message
	when in verbose mode, for cases where we fail to find the
	alternate debug info.
	* tools/fedabipkgdiff (generate_comparison_halves): Always provide
	all associated debuginfo packages to abipkgdiff.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>