Bug 29345 - abipkgdiff is confused by symlinked binaries in RPMs
Note that this issue can be reproduced by doing:

    $ fedabipkgdiff  --self-compare -a --from fc37 dx

When a binary is a symlink to another one,
create_maps_of_package_content doesn't realize it and considers the
two binaries to be different.

Later, when comes time to self compare both binaries, their abixml
files (which are the same, by virtue of symlinks) might be written at
the same time in different threads, creating a race condition, leading
to corruption of the abixml.

This patch fixes this by teaching create_maps_of_package_content to
resolve symlinks so that it can detect when two files actually point
to the same file.

	* tools/abipkgdiff.cc (create_maps_of_package_content): Resolve
	symlinks when mapping binaries.  Don't map a binary that has
	already been seen.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>