dwarf-reader: Support DW_OP_GNU_variable_value
This solves a crash that happened with self-comparing the package
'aws' in Fedora by doing:

    $ fedabipkgdiff  --self-compare -a --from fc37 aws

When evaluating a DWARF expression with
eval_last_constant_dwarf_sub_expr, indirectly called from
die_member_offset, the DW_OP_GNU_variable_value appears not being
supported.

This patch adds the support for that.

To help with figuring that kind of issue in the future, I
have added a few asserts in the code of op_is_arith_logic.

	* src/abg-dwarf-reader.cc (op_pushes_non_constant_value): Support
	DW_OP_GNU_variable_value.
	(op_is_arith_logic): Add a number of asserts and guards here.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>