Bug 29339 - elf-helpers: Don't crash on unexpected ELF file
When get_soname_of_elf_file is given an unexpected ELF file (e.g, a
DWARF file that is at the wrong place in an RPM, for instance) it hits
an assert and aborts.  Ooops.

This patch removes the offending assert from get_soname_of_elf_file.

Note that to reproduce the initial issue one has to type:

  $ fedabipkgdiff --self-compare -a --from fc36 julia

	* src/abg-elf-helpers.cc (get_soname_of_elf_file): If the program
	header we are looking at is not what we expect, just skip it; do
	not abort.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>